fix(lib): Fix the link checking when creating symlinks
Use readlink instead of realpath when checking if a symlink is pointed
at the right location.

Realpath will canonicalize the symlink, checking the actual location it
resolves to, while readlink just gets where the link immediately points
to.
This means that if a symlink is created that points to a file in the
profile, it will fail on each check, as realpath will resolve the link's
target to the path in the store, rather to the file within the current
profile directory.