fix(games/ffxiv): Fix syncing the FFXIV WINE build
Rsync by default checks timestamps to determine what to sync.
Since the FFXIV wine is built with Nix, which sets build times to the
UNIX epoch, this means rsync will always assume no syncing is needed.

Add a flag to make rsync actually check checksums for files when
determining what to sync, to prevent it silently not syncing anything on
updates.