feat(nixos/games/steam): Add Steam configuration
Configure Steam through NixOS, since the NixOS steam module also sets up
controller and hardware udev rules for Steam, and ensures that the
appropriate OpenGL/Vulkan drivers are available in the right places to
be picked up by Steam's pressure vessel.

This is needed for pressure vessel to actually work properly with games,
and without it, parts of Steam, and newer Proton versions, don't work
properly.