fix(games/tf): Patch a regression in tinyfugue
ingwarsw/tinyfugue#26 introduced a regression by shifting the background
color flag to overlap with the bottom bit of the foreground color.

This should hopefully patch it to instead move the italics flag, albeit
to a very awkward and out-of-place position in the bitmask where it may
or may not cause issues.