feat(overlay)!: Remove unneeded overlayed packages.
Remove modifying gnupg, as the options that are enabled in the overlay
are already enabled by default, and stop modifying git to disable a test
that fails on ZFS, as it's being built on a tmpfs anyway.

This not only allows using binary caches for gnupg and git, but also for
everything that depends on them directly or indirectly, such as cargo,
and thus most things that use rust or libraries written in rust, such as
librsvg and thus libreoffice.