fix(games): Revert to lutris wine 6.5
For some reason I can't figure out, I can't get more modern versions of
wine to run properly.
Additionally, lutris wine 6.5 doesn't build with the latest GCC due to
not linking libm explicitly in several places it needs to.

So revert to lutris wine 6.5 and build it with nixos-stable for now.