revert: "feat(nixos/nightmare): Add a minimal recovery ISO"
Remove the recovery ISO from GRUB since at least in my case my boot
partitions are using LUKS, and there is no easy way for a generic ISO's
stage 1 to scan and locate its stage 2 on a LUKS partition without
adding in the UUID of the boot partitions.

Keep it in the commit history for future reference, though.

Refs: 4c8d7dd18d8fcb7d37323ebc03ed4194adc6da6d