fix(gui/display): Fix rounding issues with opacity keybindings
Due to rounding in the conversion picom-trans uses between its 0-100
scale and the actual 32-bit opacity value in X11, changing opacity with
picom-trans tends to round down the opacity by one in many cases.

transset uses doubles, preventing this rounding issue to the same
degree, but picom seems to exhibit weird behavior when it removes the
opacity property if setting a window to 100% opacity, so simply remove
that for now.